✨ Extract components from Sidebar
